LayoutTests/fast/forms/label/labels-add-htmlFor-label-expected.txt

Test that the labels list of a form control is dynamically updated when adding a label whose htmlFor points to that control.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS successfullyParsed is true

TEST COMPLETE
